Fixes as per code review comments.
