+++ Testing/functional/tests/cypress/integration/e2e/vaccineCard/authenticated.js
@@ -1,0 +1,70 @@
+const { AuthMethod, localDevUri } = require("../../../support/constants");
+
+const monthNames = [
+    "January",
+    "February",
+    "March",
+    "April",
+    "May",
+    "June",
+    "July",
+    "August",
+    "September",
+    "October",
+    "November",
+    "December",
+];
+
+const dobYearSelector =
+    "[data-testid=dateOfBirthInput] [data-testid=formSelectYear]";
+const dobMonthSelector =
+    "[data-testid=dateOfBirthInput] [data-testid=formSelectMonth]";
+const dobDaySelector =
+    "[data-testid=dateOfBirthInput] [data-testid=formSelectDay]";
+const dovYearSelector =
+    "[data-testid=dateOfVaccineInput] [data-testid=formSelectYear]";
+const dovMonthSelector =
+    "[data-testid=dateOfVaccineInput] [data-testid=formSelectMonth]";
+const dovDaySelector =
+    "[data-testid=dateOfVaccineInput] [data-testid=formSelectDay]";
+
+const feedbackPhnMustBeValidSelector = "[data-testid=feedbackPhnMustBeValid]";
+const feedbackPhnIsRequiredSelector = "[data-testid=feedbackPhnIsRequired]";
+const feedbackDobIsRequiredSelector = "[data-testid=feedbackDobIsRequired]";
+const feedbackDovIsRequiredSelector = "[data-testid=feedbackDovIsRequired]";
+
+const homeUrl = "/";
+describe("Authenticated User - Vaccine Card Page", () => {
+    it("Authenticated User - Vaccination Card - Registered Keycloak authenticated user", () => {
+        cy.enableModules([
+            "Immunization",
+            "VaccinationStatus",
+            "VaccinationStatusPdf",
+        ]);
+        cy.login(
+            Cypress.env("keycloak.username"),
+            Cypress.env("keycloak.password"),
+            AuthMethod.KeyCloak,
+            "/covid19"
+        );
+
+        cy.get("[data-testid=formTitleVaccineCard]").should("be.visible");
+        cy.get("[data-testid=statusPartiallyVaccinated]").should("be.visible");
+        cy.get("[data-testid=dose-1]").should("be.visible");
+        cy.get("[data-testid=dose-2]").scrollIntoView().should("be.visible");
+    });
+
+    it("Authenticated User - Vaccination Card - Unregistered Keycloak authenticated user", () => {
+        cy.enableModules(["Immunization", "VaccinationStatus"]);
+        cy.login(
+            Cypress.env("keycloak.unregistered.username"),
+            Cypress.env("keycloak.password"),
+            AuthMethod.KeyCloak,
+            "/covid19"
+        );
+
+        cy.get("[data-testid=noCovidImmunizationsText]")
+            .should("be.visible")
+            .contains("No records found");
+    });
+});
